# docs/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "SMS Gateway Platform",
    "description": "Enterprise Android SMS Gateway — API Documentation",
    "version": "1.0.0"
  },
  "servers": [
    { "url": "http://localhost:3000", "description": "Local development" }
  ],
  "components": {
    "securitySchemes": {
      "x-device-key": {
        "type": "apiKey",
        "name": "x-device-key",
        "in": "header",
        "description": "Device API key for authentication"
      }
    },
    "schemas": {
      "ApiResponse": {
        "type": "object",
        "properties": {
          "success": { "type": "boolean" },
          "data": {},
          "message": { "type": "string" },
          "timestamp": { "type": "string", "format": "date-time" }
        }
      },
      "RegisterDeviceDto": {
        "type": "object",
        "required": ["deviceId", "name"],
        "properties": {
          "deviceId": { "type": "string", "example": "device_001" },
          "name": { "type": "string", "example": "Samsung Galaxy S24" }
        }
      },
      "IngestSmsDto": {
        "type": "object",
        "required": ["deviceId", "sender", "message", "timestamp"],
        "properties": {
          "deviceId": { "type": "string", "example": "device_001" },
          "sender": { "type": "string", "example": "+447777888999" },
          "message": { "type": "string", "example": "OTP code 728192" },
          "timestamp": { "type": "integer", "example": 1712345678 }
        }
      }
    }
  },
  "paths": {
    "/api/device/register": {
      "post": {
        "tags": ["Devices"],
        "summary": "Register a new device",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/RegisterDeviceDto" }
            }
          }
        },
        "responses": {
          "201": { "description": "Device registered" },
          "409": { "description": "Device already exists" }
        }
      }
    },
    "/api/device/": {
      "get": {
        "tags": ["Devices"],
        "summary": "List all devices",
        "responses": {
          "200": { "description": "Device list" }
        }
      }
    },
    "/api/sms": {
      "post": {
        "tags": ["SMS"],
        "summary": "Ingest an SMS message",
        "security": [{ "x-device-key": [] }],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/IngestSmsDto" }
            }
          }
        },
        "responses": {
          "202": { "description": "SMS accepted" },
          "401": { "description": "Unauthorized" },
          "429": { "description": "Rate limited" }
        }
      },
      "get": {
        "tags": ["SMS"],
        "summary": "List SMS messages",
        "parameters": [
          { "name": "page", "in": "query", "schema": { "type": "integer", "default": 1 } },
          { "name": "limit", "in": "query", "schema": { "type": "integer", "default": 50 } },
          { "name": "deviceId", "in": "query", "schema": { "type": "string" } }
        ],
        "responses": {
          "200": { "description": "SMS list" }
        }
      }
    },
    "/api/stats": {
      "get": {
        "tags": ["Stats"],
        "summary": "Dashboard statistics",
        "responses": {
          "200": { "description": "Stats" }
        }
      }
    },
    "/api/health": {
      "get": {
        "tags": ["Health"],
        "summary": "Health check",
        "responses": {
          "200": { "description": "Service health" }
        }
      }
    }
  }
}